+++ contrib/active_problems.txt
@@ -1,9 +1,9 @@
 # augo generated by /home/runner/work/espresso/espresso/espresso_machine/build_package/test_examples.py
-fm_wavefront_tracker
-gravity_density
+fmm_tomography
+gravity_inversion
 magnetotelluric_1D
 pumping_test
-receiver_function
+receiver_function_inversion
 simple_regression
 slug_test
-xray_tracer
+xray_tomography
