+++ contrib/gravity_inversion/LICENCE
