+++ contrib/gravity_inversion/data/gravmodel1.txt
