+++ contrib/magnetotelluric_1D/magnetotelluric_1D.py
@@ -155,8 +155,8 @@ class Magnetotelluric1D(EspressoProblem):
     
     def plot_data(self, data1, data2 = None, Cm = None):
         nf = len(self._freqs)
-        log10_rho = data[:nf]
-        phase = data[nf:]
+        log10_rho = data1[:nf]
+        phase = data1[nf:]
         fig, axs = plt.subplots(2,1,sharex = True,figsize = (4,4), num = 2)
         if Cm is None:
             axs[0].plot(np.log10(1/self._freqs), log10_rho,'k.')
@@ -187,7 +187,7 @@ class Magnetotelluric1D(EspressoProblem):
         return fig
 
     def misfit(self, data1, data2, Cm_inv = None):
-        res = data - data2
+        res = data1 - data2
         if Cm_inv is None:
             return float(res.T @ res)
         else:
