+++ contrib/receiver_function_inversion/LICENCE
