+++ contrib/receiver_function_inversion/RFsubs/dpythag.f
