+++ contrib/receiver_function_inversion/RFsubs/four1.f
