+++ contrib/receiver_function_inversion/RFsubs/ran3.f
