+++ contrib/receiver_function_inversion/RFsubs/rfi_param.inc
