+++ contrib/receiver_function_inversion/RFsubs/theo.f
