+++ contrib/receiver_function_inversion/RFsubs/theo_noise.f
