+++ contrib/receiver_function_inversion/RFsubs/whichcell.f90
