+++ contrib/receiver_function_inversion/__init__.py
@@ -4,12 +4,12 @@ It's NOT recommended to put extra code inside this file. Code inside this file w
 executed when this submodule is imported, so adding things in this file can slow down
 the importing process of `espresso`.
 
-For contributors, add any intialisation code for your problem into receiver_function.py, 
+For contributors, add any intialisation code for your problem into receiver_function_inversion.py, 
 under the method `__init__()` of the class `ReceiverFunctionInversion`.
 
 Don't touch this file unless you know what you are doing :)
 """
 
-from .receiver_function import *
+from .receiver_function_inversion import *
 
 __all__ = [ "ReceiverFunctionInversion" ]
