+++ contrib/xray_tomography/__init__.py
@@ -5,6 +5,6 @@ Code inside this file will be executed when this submodule is imported.
 For contributors: feel free to add your initialisation code here if needed.
 """
 
-from .xray_tracer import *
+from .xray_tomography import *
 
 __all__ = [ "XrayTomography" ]
