+++ contrib/xray_tomography/data/csiro_logo.png
