+++ contrib/xray_tomography/data/example1.dat
