+++ contrib/xray_tomography/data/example2.dat
