+++ contrib/xray_tomography/data/example3.dat
