+++ contrib/xray_tomography/data/inlab_logo.png
